requirements.txt
requests
pandas
numpy
matplotlib
seaborn
scikit-learn
flask
dash
gunicorn
dash-bootstrap-components
psycopg2-binary